+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfix
 
 - Ward intercepts Legion nuclear missiles
+- The player list icon and the build bar theme recognise Legion units in a game that tags unit specs per army, such as Galactic War
 
 ## v1.32.2 - 2026-08-27